Fixed Jumping + Animator Script
Realized player height was 0 so it was always checking negatives, so I adjusted it to check from top of head and it works (for slopes too).

Created an Animator script, not sure if we'll need it as I don't know if Unity has a built in animator system that just auto queues them or what.  But incase it doesn't, this script will simply handle animation playing outside our logic script for the player.  Needs to be linked up to animations + needs a IsFinished implemented